remove erb parsing
if we do this we open the door to arbitrary ruby which is going to be
context dependent and it will eliminate the possibility of ever
autoconverting from YAML to ruby.

if we build up a YAML structure we can better ensure that there's a
static mapping of the YAML code to ruby for autoconversion.