a plan for Windows!
Based on Stuart Preston's work. Currently depends on the very large
ruby-plus-devkit package that is a repackaging of
RubyInstaller+DevKit[1]. The DevKit is essentially MSYS2, which is its
own software distribution. It can get large. It is depended on currently
because it RubyInstaller+DevKit has been the most common Ruby on Windows
solution for build environment and library ecosystem.

Whatever Ruby it depends on is expected to have its executables on the
PATH and any gems it ships with (say, bundler, rake, etc) added to
GEM_PATH which is totally a thing a package can do in SetupEnvironment
and pushing directories onto GEM_HOME.

For the gems whose code _lives_ in this repo, the build uses `gem build`
to package up the files that are "shippable" and then unpacks to the
Habitat source cache build directory. This avoids shipping extraneous
files not meant for "production".

Includes some functions for Gemfile.lock parsing shenanigans to know
which gems are coming from paths and from git repos.

[1] https://rubyinstaller.org/about/comparison/

Co-authored-by: Stuart Preston <stuart@chef.io>
Co-authored-by: John Snow <thelunaticscripter@outlook.com>
Signed-off-by: Robb Kidd <robb@thekidds.org>